first commit, base files: pythonWeb is the first creation, running the QT5 WebKit system, based off, and extended apon The Linux Voice Tutorial "buildl a web browser with 20 lines of python" https://www.linuxvoice.com/build-a-web-browser-with-20-lines-of-python/
pyWebTest is the second iteration of pythonWeb, still running the QT5 WebKit system, but wil save the offline data, and history to files, in a folder titled .pythonWeb in your
home directory, If you sync this directory with other computers also running this system, it will save the history with the hostname of the computer that you used to view that
page.

pyWebEngine.py
is the third 'testing' iteration of pythonWeb, it it running the QT5 WebKitEngine based on Chrome.
I am still testing and getting this working correctly.